+++ app/components/svg-pie-sector.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-svg-pie/components/svg-pie-sector';