+++ app/components/svg-pie-segment.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-svg-pie/components/svg-pie-segment';